+++ CHANGELOG.rst
@@ -132,3 +132,4 @@ Changelog
 * Refactored task sync class into smaller functions that make the logic easier
   to read and modify.
 * Added last_modified property to Task (and TrelloTask, HabiticaTask)
+* Added modification time check for deciding whether to update tasks